Alarm can ring now
Added icons for the app since it was necessary for the flutter notification libary

Changed AndroidMaifest, again because of the flutter notification libary

Minor change in the ui. but those can/should be talked about with all participants

Added NotificationScreen

Added ring functionality

Added necessary libaries

This is just a(could be viewed as the first) alpha version of the app. Bugs may occure.